Переписал, чтобы была возможность реализации нескольких уровней бота. Слепил logic и game
